+++ src/state/exception.py
@@ -1,6 +1,6 @@
 # Copyright 2025 Canonical Ltd.
 # See LICENSE file for licensing details.
-"""gateway-api-integrator charm base exceptions."""
+"""haproxy charm base exceptions."""
 
 
 class CharmStateValidationBaseError(Exception):
