+++ tests/integration/legacy/test_cos.py
