+++ tox.ini
@@ -43,6 +43,7 @@ deps =
     flake8-docstrings-complete>=1.0.3
     flake8-test-docs>=1.0
     isort
+    jubilant==1.1.1
     mypy
     pep8-naming
     pydocstyle>=2.10
@@ -106,7 +107,9 @@ deps =
     protobuf==3.20.3
     pydantic
     pytest
+    jubilant==1.1.1
     juju==3.6.1.0
+    pytest-jubilant==0.7.1
     pytest-operator
     pytest-asyncio
     websockets<14.0 # https://github.com/juju/python-libjuju/issues/1184
